Fix encoder math; use MotorReader in Encoder; add Shooter.stop() cleanup (safe)

+++ TeamCode/src/main/java/com/teamcode/subsystems/Odometry.java
@@ -40,7 +40,7 @@ public class Odometry {
 
     private static double ticksToInches(int ticks) {
         double wheelCircum = Math.PI * WHEEL_DIAMETER_IN;
-        return (ticks / TICKS_PER_REV) * (wheelCircum * GEAR_RATIO);
+        return (ticks / (double) TICKS_PER_REV) * (wheelCircum * GEAR_RATIO);
     }
 
     public Odometry(HardwareMap hw) {
